Add optional password authentication to the socket server (ident+dma)
The "ident" server running on UDP port 64 is used to discover Ultimate devices
on the local network. If a JSON response is sent to the client, a new boolean
field "requires_authentication" is included to indicate if future communication
on the "dma" TCP port 64 should start with a SOCKET_CMD_AUTHENTICATE (0xFF1F)
command. The argument to this new command is simply a password, and the
response to the client is single byte (0x00 for password incorrect, 0x01 for
password correct). An incorrect password will cause the Ultimate to close the
connection after a brief delay (to slow down password brute forcing).