Add optional password authentication to the Web server API
The web API under http://<ultimate-ip>/v1/ is now optionally password
protected. The password should be supplied using a 'X-Password' HTTP header.

Example use:

  curl --header "X-Password: my-secret-password" http://192.168.2.64/v1/version

To determine if a password is needed the client should do a GET on /v1/version
(without a password) and if the HTTP response status code is 401 or 403 then a
password is required.

The HTML page served by the Ultimate web server does not yet support entering a
password.